backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
pydantic==2.10.6
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
